pyproject.toml
[tool.poetry]
name = "update-project1"
version = "0.1.0"
description = "updated task sorting project"
authors = ["pytonistdarkweb <edgaravetan2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.6"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncio = "^3.4.3"
mypy = "^1.14.0"
pytest = "^8.3.4"
passlib = "^1.7.4"
asyncpg = "^0.30.0"
poetry = "^1.8.5"
pylint = "^3.3.3"
python-language-server = "^0.36.2"
uvicorn = "^0.27.0"
python-jose = "^3.3.0"
alembic = "^1.14.0"
bcrypt = "^4.2.1"
trio = "^0.28.0"
pytest-asyncio = "^0.25.1"
pytest-trio = "^0.8.0"
anyio = "^4.7.0"
pytest-anyio = "^0.0.0"
asgi-lifespan = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


